worked on confirmation popup for clearing cart
